requirements.txt
asgiref==3.4.1
astroid==2.8.0
biopython==1.79
black==21.9b0
click==8.0.1
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
django-import-export==2.6.0
django-import-export-customized==1.1.1.dev0
et-xmlfile==1.1.0
gunicorn==20.1.0
isort==5.9.3
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.9
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.1
pylint==2.11.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.24
sqlparse==0.4.2
tablib==3.0.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
whitenoise==5.3.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
Django>=3.0,<4.0
psycopg2>=2.8
Pillow==8.4.0